a little refactor
