+++ go_back/internal/handlers/item_handler.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"gng/internal/services"
-	"gng/internal/utils"
+	"gng/internal/utils/logger"
 
 	"github.com/go-chi/chi/v5"
 )
