+++ go_back/internal/helper/env.go
@@ -1,0 +1,12 @@
+package helper
+
+import (
+	"os"
+)
+
+func GetEnv(key, fallback string) string {
+	if value := os.Getenv(key); value != "" {
+		return value
+	}
+	return fallback
+}
